+++ app/application/static/.mapbox_token
