+++ app/application/static/images/fig2.png
