+++ app/application/static/palette.png
