+++ app/modeling/farsite.py
@@ -9,8 +9,8 @@
 ################################################
 
 # weather processing module (thank you nathan)
-from exploration.data.current_weather import CurrentWeather
-from modeling.rothermel import compute_surface_spread
+from modeling.weather import CurrentWeather
+from app.modeling.rothermel import compute_surface_spread
 
 # Data containers and pre-processing
 import pickle
@@ -18,6 +18,7 @@ import pickle
 # Computational Tools
 import numpy as np
 import pandas as pd
+import xarray as xr
 import os
 
 
@@ -64,6 +65,7 @@ def regrid(AFC, INPUT, wind_speed, wind_dir, new_i, new_j, new_x, new_y, cell):
 
     return new_x, new_y
 
+
 def handle_new_fire_point(new_frontier, FIRES, NB, AFC, PIFC, INPUT, FUEL, wind_speed, wind_dir, cell, new_i, new_j,
                           new_x, new_y):
     """
@@ -108,7 +110,71 @@ def handle_new_fire_point(new_frontier, FIRES, NB, AFC, PIFC, INPUT, FUEL, wind_
             FIRES.add((new_i, new_j))
 
 
-def pre_burn(lat, lon, path_pickle):
+def prepare_data(path_farsite="data/farsite.nc", path_fueldict="data/FUEL_DIC.csv"):
+    """
+    Prepares the data required for fire modeling
+    :param lat: latitude coordinate of ignition point
+    :param lon: longitude coordinate of ignition point
+    :param path_farsite: path to the file farsite.nc, containing LANDFIRE data
+    :param path_fueldict: path to the file FUEL_DIC.csv, containing translation info for fuel types
+    :return: INPUT and FUEL, arrays described below, X and Y arrays of lat/lon coordinates
+    """
+    # we also need grid coordinates for the lat/lon, so lets perform that conversion
+
+    LATLON = xr.open_dataset(path_farsite, decode_coords="all")
+    LATLON = LATLON.rio.reproject("EPSG:4326")
+    X = LATLON["x"].data
+    Y = LATLON["y"].data
+
+    # # #
+    # Fuel data processing
+    # # #
+
+    LANDFIRE = xr.open_dataset(path_farsite, decode_coords="all")
+    FUEL = LANDFIRE['US_210F40'][:].data
+    ELEV = LANDFIRE['US_DEM'][:].data
+
+    INPUT = np.zeros((FUEL.shape[0], FUEL.shape[1], 6), dtype=np.float32)  # 32 bit float for efficiency
+
+    # from fuel types we need:
+    #
+    # (dim 0) Fuel Bed Depth (delta)  - Mean fuel array value in ft
+    # (dim 1) SA to Vol Ratio (sigma) - (ft^-1)
+    # (dim 2) Oven-dry fuel load (w_0)- (lb/ft^2) must convert from tons/acre
+    # (dim 3) Extinction Moisture (Mx)- Should be very close to fuel moisture
+    # (dim 4) Fuel Moisture (Mf)      - Approximated as proportion of extinction moisture
+    #
+    #
+    # from elevation we need:
+    #
+    # (dim 5) Elevation in meters
+
+    FUEL_TYPE_MAP = pd.read_csv(path_fueldict, header='infer').set_index('VALUE')
+    FUEL_TYPE_MAP = {float(ind): np.array([FUEL_TYPE_MAP['FuelBedDepth'][ind],
+                                           FUEL_TYPE_MAP['SAV'][ind],
+                                           FUEL_TYPE_MAP['OvenDryLoad'][ind],
+                                           FUEL_TYPE_MAP['Mx'][ind] / 100,
+                                           (FUEL_TYPE_MAP['Mx'][ind] * .95) / 100, 0])
+                     for ind in FUEL_TYPE_MAP.index}
+
+    for i in range(INPUT.shape[0]):
+        for j in range(INPUT.shape[1]):
+            # pull essential data
+            INPUT[i, j, :] = FUEL_TYPE_MAP[FUEL[i, j] if FUEL[i, j] else -9999.]
+
+            # (dim 2): ton/acre -> lb/ft^2
+            INPUT[i, j, 2] *= .0459137
+
+            # get elevation for final dimension
+            INPUT[i, j, 5] = ELEV[i,j]
+
+    data = INPUT, FUEL, X, Y
+    with open("data/farsite.pickle", "wb") as f:
+        pickle.dump(data, f, protocol=pickle.HIGHEST_PROTOCOL)
+    return data
+
+
+def pre_burn(lat, lon, path_pickle="data/farsite.pickle"):
     """
     Processes a provided data pickle, as well as lat/lon to get info for burn
     :param lat: latitudinal coordinate of ignition
@@ -116,11 +182,14 @@ def pre_burn(lat, lon, path_pickle):
     :param path_pickle: path to the preprocessed pickle data
     :return: unpickled data, istart, jstart, wind speed, wind direction
     """
+    if os.path.exists(path_pickle):
+        with open(path_pickle, "rb") as f:
+            data = pickle.load(f)
+    else:
+        data = prepare_data()
+
     # INPUT (landfire stuff), FUEL (raw fuel type), X (longitudes), Y (latitudes)
     # INPUT must be expanded to account for slope in direction of wind
-    with open(path_pickle, "rb") as f:
-        data = pickle.load(f)
-
     # get starting cell
     i_start, j_start = np.argmin(np.abs(data[2] - lon)), np.argmin(np.abs(data[3] - lat))
 
@@ -185,13 +254,11 @@ def pre_burn(lat, lon, path_pickle):
     return pre_burn_data
 
 
-def burn(lat, lon, path_farsite="data/farsite.nc", path_fueldict="data/FUEL_DIC.csv", path_pickle="data/farsite.pickle", mins=50):
+def burn(lat, lon, path_pickle="data/farsite.pickle", mins=50):
     """
     Burning down the house
     :param lat: latitude of ignition
     :param lon: longitude of ignition
-    :param path_farsite: path to `farsite.nc`
-    :param path_fueldict: path to `FUEL_DIC.csv`
     :param path_pickle: path to preprocessed pickle data
     :param mins: number of one minute iterations to burn for
     :return: A set of cells burned after all iterations
