+++ exploration/modeling/farsite_draft.py
